Add a common file for functions that are basically the same in both scene and history ingest
